fix(law-kb): ship onnxruntime-node in packaged app for vector search
- bundle.mjs: keep onnxruntime-node external (no stub plugin), add CJS
  banner so require() resolves at runtime from resources/node_modules/
- electron-builder.yml: copy onnxruntime-node dist+bin+package.json to
  resources/node_modules/onnxruntime-node/ so MCP subprocess finds it
  via Node.js directory traversal from resources/mcp-servers/
- embedder.ts: return null on runtime failure; callers fall back to FTS
- search.ts: null guards after getEmbedding() in all three search fns
- indexer.ts: add if (!emb) continue; null guards (TS18047 fix)

+++ packages/law-kb-mcp/src/__tests__/embedder.test.ts
@@ -20,7 +20,7 @@ describe('getEmbedding', () => {
   it('returns Float32Array of length 1024', async () => {
     const result = await getEmbedding('劳动合同解除');
     expect(result).toBeInstanceOf(Float32Array);
-    expect(result.length).toBe(1024);
+    expect((result as Float32Array).length).toBe(1024);
   });
 
   it('accepts mode parameter without error', async () => {
